# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: WealthRM Order Management API
  description: |
    Comprehensive API for managing mutual fund orders, portfolios, goals, and client relationships.
    
    This API provides endpoints for:
    - Order management (purchase, redemption, switch)
    - Portfolio tracking and analysis
    - Goal-based investing
    - Client and prospect management
    - SIP (Systematic Investment Plans)
    - Quick orders and favorites
    - Webhooks for real-time notifications
    - Bulk order processing
    
    ## Authentication
    All endpoints require session-based authentication. Users must be logged in via `/api/auth/login`.
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse. Standard rate limits:
    - 100 requests per minute per user
    - 1000 requests per hour per user
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "message": "Error description",
      "errors": ["Detailed error messages"],
      "code": "ERROR_CODE"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@wealthrm.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.wealthrm.com/api
    description: Production server
...
paths:
...
  /order-management/orders:
    get:
      tags:
        - Orders
      summary: Get orders
      description: Retrieve orders with optional filtering
      operationId: getOrders
      security:
        - sessionAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by order status
          required: false
          schema:
            type: string
            enum: [Pending, "Pending Approval", "In Progress", "Settlement Pending", Executed, Settled, Failed, "Settlement Reversal", Cancelled]
        - name: startDate
          in: query
          description: Start date filter (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date filter (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Orders
      summary: Submit order
      description: Submit a new order for processing
      operationId: submitOrder
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderFormData'
      responses:
        '201':
          description: Order submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /order-management/orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieve a specific order by its ID
      operationId: getOrderById
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /bulk-orders:
    post:
      tags:
        - Bulk Orders
      summary: Submit bulk orders
      description: Submit multiple orders in a single request
      operationId: submitBulkOrders
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkOrderRequest'
      responses:
        '202':
          description: Bulk order submitted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOrderResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bulk-orders/{batchId}:
    get:
      tags:
        - Bulk Orders
      summary: Get bulk order status
      description: Get the status of a bulk order batch
      operationId: getBulkOrderStatus
      security:
        - sessionAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bulk order batch status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOrderStatus'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie authentication
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
        modelOrderId:
          type: string
          example: MO-20241215-ABC12
        clientId:
          type: integer
        orderFormData:
          $ref: '#/components/schemas/OrderFormData'
        status:
          type: string
          enum: [Pending, "Pending Approval", "In Progress", "Settlement Pending", Executed, Settled, Failed, "Settlement Reversal", Cancelled]
        submittedAt:
          type: string
          format: date-time
        authorizedAt:
          type: string
          format: date-time
        authorizedBy:
          type: integer
        rejectedAt:
          type: string
          format: date-time
        rejectedReason:
          type: string
        rtaRefNo:
          type: string
        orderPaymentReference:
          type: string
        pgPaymentReferenceNo:
          type: string
        paymentlinkedstatus:
          type: string
        payremarks:
          type: string
        ipAddress:
          type: string
        traceId:
          type: string
    OrderFormData:
      type: object
      required:
        - cartItems
        - transactionMode
      properties:
        cartItems:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        transactionMode:
          $ref: '#/components/schemas/TransactionModeData'
        investmentAccount:
          $ref: '#/components/schemas/InvestmentAccount'
        settlementAccount:
          $ref: '#/components/schemas/InvestmentAccount'
        branchCode:
          type: string
        nominees:
          type: array
          items:
            $ref: '#/components/schemas/Nominee'
        optOutOfNomination:
          type: boolean
        dividendInstruction:
          type: string
        euin:
          type: string
        fullSwitchData:
          $ref: '#/components/schemas/FullSwitchData'
        fullRedemptionData:
          $ref: '#/components/schemas/FullRedemptionData'
    CartItem:
      type: object
      required:
        - id
        - productId
        - schemeName
        - transactionType
        - amount
      properties:
        id:
          type: string
        productId:
          type: integer
        schemeName:
          type: string
        transactionType:
          type: string
          enum: [Purchase, Redemption, Switch, "Full Redemption", "Full Switch"]
        amount:
          type: number
        units:
          type: number
        nav:
          type: number
        settlementAccount:
          type: string
        branchCode:
          type: string
        mode:
          type: string
        dividendInstruction:
          type: string
        closeAc:
          type: boolean
        orderType:
          type: string
          enum: ["Initial Purchase", "Additional Purchase"]
        sourceSchemeId:
          type: integer
        sourceSchemeName:
          type: string
    TransactionModeData:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum: [Physical, Email, Telephone]
        email:
          type: string
        phoneNumber:
          type: string
        physicalAddress:
          type: string
        euin:
          type: string
    InvestmentAccount:
      type: object
      properties:
        id:
          type: integer
        accountNumber:
          type: string
        bankName:
          type: string
        ifscCode:
          type: string
        accountType:
          type: string
        isActive:
          type: boolean
    Nominee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        relationship:
          type: string
        dateOfBirth:
          type: string
          format: date
        pan:
          type: string
        percentage:
          type: number
        guardianName:
          type: string
        guardianPan:
          type: string
        guardianRelationship:
          type: string
        isMinor:
          type: boolean
    FullSwitchData:
      type: object
      properties:
        sourceScheme:
          type: string
        targetScheme:
          type: string
        units:
          type: number
        closeAc:
          type: boolean
    FullRedemptionData:
      type: object
      properties:
        schemeName:
          type: string
        units:
          type: number
        amount:
          type: number
        closeAc:
          type: boolean
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Order'
    BulkOrderRequest:
      type: object
      required:
        - orders
      properties:
        orders:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/OrderFormData'
        options:
          type: object
          properties:
            stopOnError:
              type: boolean
              description: Stop processing if any order fails
              default: false
            validateOnly:
              type: boolean
              description: Only validate orders without submitting
              default: false
    BulkOrderResponse:
      type: object
      properties:
        success:
          type: boolean
        batchId:
          type: string
        message:
          type: string
        summary:
          type: object
          properties:
            total:
              type: integer
            submitted:
              type: integer
            failed:
              type: integer
            errors:
              type: array
              items:
                type: object
                properties:
                  index:
                    type: integer
                  error:
                    type: string
    BulkOrderStatus:
      type: object
      properties:
        batchId:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, partial]
        total:
          type: integer
        processed:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        errors:
          type: array
          items:
            type: object
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: array
          items:
            type: string
        code:
          type: string
        warnings:
          type: array
          items:
            type: string